+++ app/src/main/java/com/hw/myapplication/callbacks/CallBack_MovePlayer.java
@@ -2,5 +2,5 @@ package com.hw.myapplication.callbacks;
 
 public interface CallBack_MovePlayer {
     void movePlayer(int direction);
-    void gameSpeed(int speed);
+    void gameSpeed(long speed);
 }
